fix: incorrect connections / phasediff tests
